+++ lib/feature/inventory/inventory_by_nom/models/cells_by_nom.dart
@@ -28,7 +28,7 @@ class CellsByNom {
 class Cell {
   final String code;
   final String name;
-  final int planCount;
+  final double planCount;
   final int factCount;
   final String nomStatus;
