+++ lib/feature/return/return_client/models/return_nom_dto.dart
@@ -25,19 +25,17 @@ class ReturnNomDTO {
   final List<BarcodeDTO> barcodes;
   final double? qty;
   final double? count;
-    @JsonKey(name: 'nom_status')
-
+  @JsonKey(name: 'nom_status')
   final String? nomStatus;
 
-  ReturnNomDTO({
-    required this.name,
-    required this.article,
-    required this.barcodes,
-    required this.number,
-    required this.qty,
-    required this.count,
-    required this.nomStatus
-  });
+  ReturnNomDTO(
+      {required this.name,
+      required this.article,
+      required this.barcodes,
+      required this.number,
+      required this.qty,
+      required this.count,
+      required this.nomStatus});
 
   factory ReturnNomDTO.fromJson(Map<String, dynamic> json) =>
       _$ReturnNomDTOFromJson(json);
@@ -46,7 +44,7 @@ class ReturnNomDTO {
 @JsonSerializable(createPerFieldToJson: false)
 class BarcodeDTO {
   final String? barcode;
-  final int? ratio;
+  final double? ratio;
 
   BarcodeDTO({required this.barcode, required this.ratio});
 
@@ -77,4 +75,3 @@ class CellDTO {
   factory CellDTO.fromJson(Map<String, dynamic> json) =>
       _$CellDTOFromJson(json);
 }
-
